requirements_streamlit.txt
git commit -m "Switch to # Alternative requirements for maximum compatibility
streamlit==1.28.1
tensorflow==2.12.0
numpy==1.23.5
pillow==9.5.0
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
